requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
rembg==2.0.65
onnxruntime==1.21.0
Pillow==11.1.0
python-multipart==0.0.20
